pyproject.toml
[project]
name = "technical-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "finnhub-python>=2.4.23",
    "google-generativeai>=0.8.5",
    "langchain-google-genai>=2.0.10",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "uvicorn>=0.34.2",
]
